+++ example/src/pages/403.react.js
@@ -2,18 +2,12 @@
 
 import * as React from "react";
 
-import { DefaultErrorPage } from "tabler-react";
+import { Error403Page } from "tabler-react";
 
 type Props = {||};
 
 function Error403(props: Props): React.Node {
-  return (
-    <DefaultErrorPage
-      title="403"
-      subtitle="Oops... You just found an error page..."
-      details="We are sorry but you do not have permission to access this page..."
-    />
-  );
+  return <Error403Page />;
 }
 
 export default Error403;