+++ example/src/pages/500.react.js
@@ -2,18 +2,12 @@
 
 import * as React from "react";
 
-import { DefaultErrorPage } from "tabler-react";
+import { Error500Page } from "tabler-react";
 
 type Props = {||};
 
 function Error500(props: Props): React.Node {
-  return (
-    <DefaultErrorPage
-      title="500"
-      subtitle="Oops... You just found an error page..."
-      details="We are sorry but your request contains bad syntax and cannot be fulfilled..."
-    />
-  );
+  return <Error500Page />;
 }
 
 export default Error500;