+++ CHANGELOG.md
@@ -1,1 +1,20 @@
+<a name="1.5.1"></a>
 
+## [1.5.1](https://github.com/tabler/tabler-react/compare/v1.5.0...v1.5.1) (2018-05-25)
+
+### Bug Fixes
+
+* **semantic-release:** fix config ([5ba703d](https://github.com/tabler/tabler-react/commit/5ba703d))
+
+<a name="1.5.0"></a>
+
+# [1.5.0](https://github.com/tabler/tabler-react/compare/v1.4.0...v1.5.0) (2018-05-25)
+
+### Bug Fixes
+
+* **tabler css:** copy tabler css before build and start ([2062c71](https://github.com/tabler/tabler-react/commit/2062c71))
+
+### Features
+
+* **Form:** Add form and input event props ([c109724](https://github.com/tabler/tabler-react/commit/c109724))
+* **FormInput:** Add error prop ([2d98e75](https://github.com/tabler/tabler-react/commit/2d98e75))